Removed signup wall
At first I tried deleting the cookies that keep track of the signup wall, but that started causing other errors on the website, so instead I deleted the html element whenever it appears. However, the signup wall appears quite frequently, so I kept the cookie code in and I might go back to it.