adds comments of original poster
I swear to fucking god queryselectors are the most annoying shit in the world. With the other stuff in javascript, you can use intellisense and linter errors to figure out what you can and can't do, but with queryselectors it's a giant magic string that either works 70% of the time or gives you an error that tells you your queryselector is wrong without telling you why.
Anyway there's a bug with comments that have a "(more)" button, but once I fix that the bulk of this extension is officially done.